Added Holdings menu

+++ TopStocks/Models/Price.cs
@@ -2,8 +2,8 @@
 {
     public class Price
     {
-        public int CurrentPrice { get; set; }
-        public int WeekHighPrice { get; set; }
-        public int WeekLowPrice { get; set; }
+        public float CurrentPrice { get; set; }
+        public float WeekHighPrice { get; set; }
+        public float WeekLowPrice { get; set; }
     }
 }
